chore(command): apply review comments on the regression test
- extend the test to unix (linux and darwin)
- make all the workers start their command at the same time, using a
  start barrier
- make the command output multiple lines, and check all of them are
  received in order
